interpret_as_integer refactored
