+++ logs/execution.log
@@ -1,37 +1,37 @@
-[2025-09-01 11:23:37]
+[2025-09-01 11:44:13]
 [SET-VARIABLE]
 NAME: 
  __payload__
 
 VALUE: 
- 
+ I would say that the result is negative.
 ===============END LOG ENTRY================
 
-[2025-09-01 11:23:37]
+[2025-09-01 11:44:13]
 [PIPELINE START]
-Send Prompt Example
+interpret as bool
 ===============END LOG ENTRY================
 
-[2025-09-01 11:23:46]
+[2025-09-01 11:44:16]
 [SET-VARIABLE]
 NAME: 
- response
+ as_bool_result
 
 VALUE: 
- Paris! The glittery baguette HQ where the Eiffel Tower moonlights as a giant metal hat rack for clouds.
+ False
 ===============END LOG ENTRY================
 
-[2025-09-01 11:23:46]
+[2025-09-01 11:44:16]
 [SET-VARIABLE]
 NAME: 
  __payload__
 
 VALUE: 
- What is the capital of France?
+ False
 ===============END LOG ENTRY================
 
-[2025-09-01 11:23:46]
+[2025-09-01 11:44:16]
 [PIPELINE END]
-Send Prompt Example
+interpret as bool
 ===============END LOG ENTRY================
